A stupid bug in BPU
Figured that need to give two btbresponses from Branch predictor as even if an instruction is a Branch and is predicted as NotTaken, I need to give a response so that stage3 can train it properly.
This also brought changes in stage1 to update the btbresponse for the instructions properly.